rename the teaserpage to Teaser
