+++ ecorise-react-app/src/App.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-import Teaser from './TeaserPage';
+import Teaser from './Teaser';
 const App = () => {
   return (
     <BrowserRouter>
