+++ ecorise-react-app/src/Teaser/index.js
