+++ ecorise-react-app/src/Teaser/index.test.js
